+++ src/main/java/com/qimiaochong/service/UserService.java
@@ -4,5 +4,5 @@ import com.qimiaochong.entity.User;
 
 public interface UserService {
 
-    User queryUser(Integer id);
+    User queryUser(Long id);
 }
