+++ src/main/java/com/qimiaochong/service/impl/TopicServiceImpl.java
@@ -24,7 +24,7 @@ public class TopicServiceImpl implements TopicService {
     }
 
     @Override
-    public Topic findOne(Integer id) {
+    public Topic findOne(Long id) {
         return topicDao.findById(id);
     }
 
@@ -39,22 +39,22 @@ public class TopicServiceImpl implements TopicService {
     }
 
     @Override
-    public boolean ban(Integer id) {
+    public boolean ban(Long id) {
         return topicDao.ban(id,new Date(),TopicStatus.BAN_STATUS)==1;
     }
 
     @Override
-    public boolean unBan(Integer id) {
+    public boolean unBan(Long id) {
         return topicDao.ban(id,new Date(),TopicStatus.NORMAL_STATUS)==1;
     }
 
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Long id) {
         return topicDao.delete(id,new Date(), TopicStatus.DELETE_STATUS)==1;
     }
 
     @Override
-    public boolean destroy(Integer id) {
+    public boolean destroy(Long id) {
         return topicDao.destroy(id)==1;
     }
 }
